+++ CHANGELOG.md
@@ -48,6 +48,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Security
 
 
+## [10.3.3] -- 2020-08-18
+
+### Fixed
+  * Fixed nested transaction not having the authorized user attached or context id in the update fiunction by: 
+    - add support for nested transactions in the BaseSequelize.newTransaction() call and use in the update function
+    - correctly find the outerMost transaction when publishing changes int eh sequelize hook so that nested transaction changes are ONLY published when the outer most transaction is committed
+
+
 ## [10.3.2] -- 2020-08-05
 
 ### Fixed